+++ src/lang/lang_types.mli
@@ -50,6 +50,7 @@ and descr =
   | Ground of ground
   | List of t
   | Tuple of t list
+  | Nullable of t
   | Meth of string * scheme * t
   | Zero
   | Succ of t